+++ tests/parse_from_file_spec.rs
@@ -1,0 +1,17 @@
+use std::path::PathBuf;
+
+use rstest::rstest;
+
+use beancount_parser::BeancountFile;
+
+#[rstest]
+#[case("comments.beancount", 0)]
+#[case("simple.beancount", 13)]
+#[case("official.beancount", 2151)]
+fn can_parse_example_files(#[case] file_name: &str, #[case] expected_directive_count: usize) {
+    let mut path: PathBuf = "./tests/samples".into();
+    path.push(file_name);
+    let mut file = BeancountFile::<f64>::default();
+    beancount_parser::read_files([path], |entry| file.extend(Some(entry))).unwrap();
+    assert_eq!(file.directives.len(), expected_directive_count);
+}