+++ CHANGELOG.md
@@ -7,6 +7,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+
+## [2.0.0-beta.2] - 2023-07-08
+
 ### Added
 
 * `Entry` enum which is one of `Directive`, `Option`, `Include`
@@ -424,7 +427,8 @@ Announce new status of the project in readme: The project is now in "maintenance
  - `Parser` iterator type
 
 
-[Unreleased]: https://github.com/jcornaz/beancount-parser/compare/v2.0.0-beta.1...HEAD
+[Unreleased]: https://github.com/jcornaz/beancount-parser/compare/v2.0.0-beta.2...HEAD
+[2.0.0-beta.2]: https://github.com/jcornaz/beancount-parser/compare/v2.0.0-beta.1...v2.0.0-beta.2
 [2.0.0-beta.1]: https://github.com/jcornaz/beancount-parser/compare/v1...v2.0.0-beta.1
 [1.16.2]: https://github.com/jcornaz/beancount-parser/compare/v1.16.1...v1.16.2
 [1.16.1]: https://github.com/jcornaz/beancount-parser/compare/v1.16.0...v1.16.1